Refactor: Moved DetectionListUserClient to its separate module
